--- src/types/middleware.ts
@@ -1,8 +1,0 @@
-export interface AuthMiddleware {
-  begin: () => RequestHandler;
-  callback: () => RequestHandler;
-}
-export type ValidateAuthenticatedSessionMiddleware = () => RequestHandler;
-export type EnsureInstalledMiddleware = () => RequestHandler;
-export type CspHeadersMiddleware = () => RequestHandler;
-export type RedirectToShopifyOrAppRootMiddleware = () => RequestHandler;
